New compact implementation when targeting browsers
New compact implementation should significantly reduce generated code
when targeting browsers. Instead of using two different code paths for
size calculation and serialization, we will produce intermediate data
structure in generated code and serialize it with a simple predefined
function.

It probably will be slower in microbenchmarks, but depending on the
use cases it maybe even faster in some situations because
serialization is using single code path, and it will be immediately
JIT compiled. And it is still faster than JSON serialization.